+++ Packages/com.chark.game-management/Runtime/Assets/ResourceLoader.cs
@@ -1,5 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
+using System.IO;
+using System.Threading;
+using System.Threading.Tasks;
+using CHARK.GameManagement.Serialization;
 using UnityEngine;
 using Object = UnityEngine.Object;
 
@@ -7,7 +10,14 @@ namespace CHARK.GameManagement.Assets
 {
     internal sealed class ResourceLoader : IResourceLoader
     {
-        public IEnumerable<TResource> LoadResources<TResource>(string path = null) where TResource : Object
+        private readonly ISerializer serializer;
+
+        public ResourceLoader(ISerializer serializer)
+        {
+            this.serializer = serializer;
+        }
+
+        public IEnumerable<TResource> GetResources<TResource>(string path = null) where TResource : Object
         {
             if (string.IsNullOrWhiteSpace(path))
             {
@@ -17,20 +27,53 @@ namespace CHARK.GameManagement.Assets
             return Resources.LoadAll<TResource>(path);
         }
 
-        public TResource LoadResource<TResource>(string path) where TResource : Object
+        public bool TryGetResource<TResource>(string path, out TResource resource) where TResource : Object
         {
             if (string.IsNullOrWhiteSpace(path))
             {
-                throw new ArgumentException($"{nameof(path)} must not be empty or null");
+                resource = default;
+                return false;
             }
 
-            var resource = Resources.Load<TResource>(path);
-            if (resource)
+            var loadedResource = Resources.Load<TResource>(path);
+            if (loadedResource)
             {
-                return resource;
+                resource = loadedResource;
+                return true;
             }
 
-            throw new Exception($"Resource at path \"{path}\" could not be loaded");
+            resource = default;
+            return false;
+        }
+
+        public async Task<TResource> GetResourceAsync<TResource>(
+            string path,
+            CancellationToken cancellationToken
+        )
+        {
+            var actualPath = Path.Combine(Application.streamingAssetsPath, path);
+
+#if UNITY_ANDROID
+            var request = UnityEngine.Networking.UnityWebRequest.Get(actualPath);
+            var operation = request.SendWebRequest();
+
+            await Cysharp.Threading.Tasks.UnityAsyncExtensions.ToUniTask(
+                operation,
+                cancellationToken: cancellationToken
+            );
+
+            var handler = request.downloadHandler;
+            var content = handler.text;
+#else
+            var content = await File.ReadAllTextAsync(actualPath, cancellationToken);
+#endif
+
+            if (serializer.TryDeserializeValue<TResource>(content, out var value))
+            {
+                return value;
+            }
+
+            return default;
         }
     }
 }