+++ Packages/com.chark.game-management/Runtime/GameManager.External.cs
@@ -10,114 +10,138 @@ namespace CHARK.GameManagement
     public abstract partial class GameManager
     {
         /// <returns>
-        /// Resources of type <see cref="TResource"/> loaded via <see cref="resourceLoader"/> at
+        /// Resources of type <see cref="TResource"/> from <see cref="resourceLoader"/> at
         /// given <paramref name="path"/>.
         /// </returns>
-        public static IEnumerable<TResource> LoadResources<TResource>(string path = null) where TResource : Object
+        public static IEnumerable<TResource> GetResources<TResource>(string path = null) where TResource : Object
         {
             var gameManager = GetGameManager();
             var resourceLoader = gameManager.resourceLoader;
 
-            return resourceLoader.LoadResources<TResource>(path);
+            return resourceLoader.GetResources<TResource>(path);
         }
 
         /// <returns>
-        /// Resource of type <see cref="TResource"/> loaded via <see cref="resourceLoader"/> at
-        /// given <paramref name="path"/>. If resource can't be loaded an <see cref="System.Exception"/>
-        /// is thrown.
+        /// <c>true</c> if resource of type <see cref="TResource"/> is retrieved from
+        /// <see cref="resourceLoader"/> at given <paramref name="path"/> or <c>false</c> otherwise.
         /// </returns>
-        public static TResource LoadResource<TResource>(string path) where TResource : Object
+        public static bool TryGetResource<TResource>(string path, out TResource resource) where TResource : Object
         {
             var gameManager = GetGameManager();
             var resourceLoader = gameManager.resourceLoader;
 
-            return resourceLoader.LoadResource<TResource>(path);
+            return resourceLoader.TryGetResource(path, out resource);
         }
 
         /// <returns>
-        /// Value retrieved from <see cref="runtimeStorage"/> at given <paramref name="key"/>
+        /// Task containing a resource of type <see cref="TResource"/> retrieved from
+        /// <see cref="resourceLoader"/>  at given <paramref name="path"/> or <c>default</c> if
+        /// value could not be retrieved.
+        /// </returns>
+        public static Task<TResource> GetResourceAsync<TResource>(string path)
+        {
+            var gameManager = GetGameManager();
+            var resourceLoader = gameManager.resourceLoader;
+
+            return resourceLoader.GetResourceAsync<TResource>(path);
+        }
+
+        /// <returns>
+        /// Value retrieved from <see cref="runtimeStorage"/> at given <paramref name="path"/>
         /// asynchronously or <c>default</c> if no value is could be retrieved.
         /// </returns>
-        public static Task<TValue> GetRuntimeValueAsync<TValue>(string key)
+        public static Task<TValue> GetRuntimeValueAsync<TValue>(string path)
         {
             var gameManager = GetGameManager();
             var runtimeStorage = gameManager.runtimeStorage;
 
-            return runtimeStorage.GetValueAsync<TValue>(key);
+            return runtimeStorage.GetValueAsync<TValue>(path);
         }
 
         /// <returns>
-        /// <c>true</c> if <paramref name="value"/> is retrieved by given <paramref name="key"/>
+        /// <c>true</c> if <paramref name="value"/> is retrieved by given <paramref name="path"/>
         /// from <see cref="runtimeStorage"/>.
         /// </returns>
-        public static bool TryGetRuntimeValue<TValue>(string key, out TValue value)
+        public static bool TryGetRuntimeValue<TValue>(string path, out TValue value)
         {
             var gameManager = GetGameManager();
             var runtimeStorage = gameManager.runtimeStorage;
 
-            return runtimeStorage.TryGetValue(key, out value);
+            return runtimeStorage.TryGetValue(path, out value);
         }
 
         /// <summary>
-        /// Persist a <paramref name="value"/> by given <paramref name="key"/> asynchronously to
+        /// Persist a <paramref name="value"/> by given <paramref name="path"/> asynchronously to
         /// <see cref="runtimeStorage"/>.
         /// </summary>
-        public static Task SetRuntimeValueAsync<TValue>(string key, TValue value)
+        public static Task SetRuntimeValueAsync<TValue>(string path, TValue value)
         {
             var gameManager = GetGameManager();
             var runtimeStorage = gameManager.runtimeStorage;
 
-            return runtimeStorage.SetValueAsync(key, value);
+            return runtimeStorage.SetValueAsync(path, value);
         }
 
         /// <summary>
-        /// Persist a <paramref name="value"/> by given <paramref name="key"/> to
+        /// Persist a <paramref name="value"/> by given <paramref name="path"/> to
         /// <see cref="runtimeStorage"/>.
         /// </summary>
-        public static void SetRuntimeValue<TValue>(string key, TValue value)
+        public static void SetRuntimeValue<TValue>(string path, TValue value)
         {
             var gameManager = GetGameManager();
             var runtimeStorage = gameManager.runtimeStorage;
 
-            runtimeStorage.SetValue(key, value);
+            runtimeStorage.SetValue(path, value);
         }
 
-        // TODO: add async variant!
         /// <summary>
-        /// Delete persisted <see cref="runtimeStorage"/>  value at given <paramref name="key"/>.
+        /// Delete persisted <see cref="runtimeStorage"/> value at given <paramref name="path"/>
+        /// asynchronously from <see cref="runtimeStorage"/>.
         /// </summary>
-        public static void DeleteRuntimeValue(string key)
+        public static Task DeleteRuntimeValueAsync(string path)
         {
             var gameManager = GetGameManager();
             var runtimeStorage = gameManager.runtimeStorage;
 
-            runtimeStorage.DeleteValue(key);
+            return runtimeStorage.DeleteValueAsync(path);
+        }
+
+        /// <summary>
+        /// Delete persisted <see cref="runtimeStorage"/> value at given <paramref name="path"/>
+        /// from <see cref="runtimeStorage"/>.
+        /// </summary>
+        public static void DeleteRuntimeValue(string path)
+        {
+            var gameManager = GetGameManager();
+            var runtimeStorage = gameManager.runtimeStorage;
+
+            runtimeStorage.DeleteValue(path);
         }
 
         /// <returns>
-        /// <c>true</c> if <paramref name="value"/> is retrieved by given <paramref name="key"/>
+        /// <c>true</c> if <paramref name="value"/> is retrieved by given <paramref name="path"/>
         /// from <see cref="EditorStorage"/>.
         /// </returns>
-        public static bool TryGetEditorValue<TValue>(string key, out TValue value)
+        public static bool TryGetEditorValue<TValue>(string path, out TValue value)
         {
-            return EditorStorage.TryGetValue(key, out value);
+            return EditorStorage.TryGetValue(path, out value);
         }
 
         /// <summary>
-        /// Persist a <paramref name="value"/> by given <paramref name="key"/> to
+        /// Persist a <paramref name="value"/> by given <paramref name="path"/> to
         /// <see cref="EditorStorage"/>.
         /// </summary>
-        public static void SetEditorValue<TValue>(string key, TValue value)
+        public static void SetEditorValue<TValue>(string path, TValue value)
         {
-            EditorStorage.SetValue(key, value);
+            EditorStorage.SetValue(path, value);
         }
 
         /// <summary>
-        /// Delete persisted <see cref="EditorStorage"/> value at given <paramref name="key"/>.
+        /// Delete persisted <see cref="EditorStorage"/> value at given <paramref name="path"/>.
         /// </summary>
-        public static void DeleteEditorValue(string key)
+        public static void DeleteEditorValue(string path)
         {
-            EditorStorage.DeleteValue(key);
+            EditorStorage.DeleteValue(path);
         }
 
         /// <returns>
@@ -188,6 +212,32 @@ namespace CHARK.GameManagement
             var messageBus = gameManager.messageBus;
 
             messageBus.RemoveListener(listener);
+        }
+
+        /// <returns>
+        /// <c>true</c> if <paramref name="serializedValue"/> is deserialized to
+        /// <paramref name="deserializedValue"/> using <see cref="serializer"/> successfully or
+        /// <c>false</c> otherwise.
+        /// </returns>
+        public static bool TryDeserializeValue<TValue>(string serializedValue, out TValue deserializedValue)
+        {
+            var gameManager = GetGameManager();
+            var serializer = gameManager.serializer;
+
+            return serializer.TryDeserializeValue(serializedValue, out deserializedValue);
+        }
+
+        /// <returns>
+        /// <c>true</c> if <paramref name="deserializedValue"/> is serialized to
+        /// <paramref name="serializedValue"/> using <see cref="serializer"/> successfully or
+        /// <c>false</c> otherwise.
+        /// </returns>
+        public static bool TrySerializeValue<TValue>(TValue deserializedValue, out string serializedValue)
+        {
+            var gameManager = GetGameManager();
+            var serializer = gameManager.serializer;
+
+            return serializer.TrySerializeValue(deserializedValue, out serializedValue);
         }
     }
 }